+++ saleor/graphql/webhook/dataloaders/__init__.py
@@ -4,19 +4,11 @@ from .models import (
     WebhooksByAppIdLoader,
     WebhooksByEventTypeLoader,
 )
-from .pregenerated_payload_for_checkout_tax import (
-    PregeneratedCheckoutTaxPayloadsByCheckoutTokenLoader,
-)
-from .pregenerated_payloads_for_checkout_filter_shipping_methods import (
-    PregeneratedCheckoutFilterShippingMethodPayloadsByCheckoutTokenLoader,
-)
-from .utils import PayloadsRequestContextByEventTypeLoader
+from .request_context import PayloadsRequestContextByEventTypeLoader
 
 __all__ = [
     "PayloadByIdLoader",
     "PayloadsRequestContextByEventTypeLoader",
-    "PregeneratedCheckoutFilterShippingMethodPayloadsByCheckoutTokenLoader",
-    "PregeneratedCheckoutTaxPayloadsByCheckoutTokenLoader",
     "WebhookEventsByWebhookIdLoader",
     "WebhooksByAppIdLoader",
     "WebhooksByEventTypeLoader",
